\name{IPSUR-package}
\alias{IPSUR-package}
\alias{IPSUR}
\alias{ipsur}
\alias{RcmdrPlugin.IPSUR}
\docType{package}
\title{
An IPSUR Plugin for the R Commander
}
\description{
  This package accompanies Introduction to Probability and Statistics Using R by G. Jay Kerns.
  It contributes functions unique to the book as well as specific configuration and selected functionality to the R Commander by John Fox.

}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.IPSUR\cr
Version: \tab 0.1-6\cr
Date: \tab 2009-12-25\cr
Depends: \tab Rcmdr (>= 1.4-0), rgl\cr
Suggests: \tab abind, distr, distrEx, e1071, effects (>= 1.0-7), foreign, grid, lattice, lmtest, MASS, mgcv, multcomp (>= 0.991-2), nlme, nnet, qcc, relimp, RODBC\cr
LazyLoad: \tab no\cr
License: \tab GPL version 2\cr
URL: \tab http://www.r-project.org, http://ipsur.r-forge.r-project.org/\cr
}

}
\author{
G. Jay Kerns <gkerns@ysu.edu> with contributions by Theophilius Boye and Tyler Drombosky, adapted from the work of John Fox et al.

Maintainer: G. Jay Kerns <gkerns@ysu.edu>
}
\keyword{ package }
