\name{RcmdrPlugin.IPSUR-package}
\alias{RcmdrPlugin.IPSUR}
\docType{package}
\title{
An IPSUR Plugin for the R Commander
}
\description{
 This package is an R Commander plugin that accompanies IPSUR, an Introduction to Probability and Statistics Using R.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.IPSUR\cr
Version: \tab 0.2-1\cr
Date: \tab 2014-09-10\cr
Imports: \tab Rcmdr (>= 2.1-0)\cr
Suggests: \tab abind, car, distr, distrEx, e1071, effects (>= 1.0-7), foreign, grid, lattice, lmtest, MASS, mgcv, multcomp (>= 0.991-2), nlme, nnet, qcc, relimp, RODBC\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org, http://ipsur.org/\cr
}

}
\author{
G. Jay Kerns <gkerns@ysu.edu> with contributions by Theophilius Boye and Tyler Drombosky, adapted from the work of John Fox et al.

Maintainer: G. Jay Kerns <gkerns@ysu.edu>
}
\keyword{ package }
